deploy-bot/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/dev.txt requirements/base.txt requirements/dev.in requirements/testing.in
#
astroid==2.2.5            # via pylint
beautifulsoup4==4.7.1     # via webtest
certifi==2019.6.16
chardet==3.0.4
click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cssselect==1.0.3          # via pyquery
django-appconf==1.0.3
django-axes==5.0.7
django-choices==1.7.0
django-debug-toolbar==2.0
django-extensions==2.1.9
django-filter==2.1.0
django-ipware==2.1.0
django-solo==1.1.3
django-webtest==1.9.6
django==2.2.2
djangorestframework-camel-case==1.0.3
djangorestframework==3.9.4
drf-nested-routers==0.91
drf-yasg==1.16.0
factory-boy==2.12.0
faker==1.0.7              # via factory-boy
freezegun==0.3.12
gemma-zds-client==0.11.0
idna==2.8
inflection==0.3.1
iso-639==0.4.5
isodate==0.6.0
isort==4.3.21
itypes==1.1.0
jinja2==2.10.1
lazy-object-proxy==1.4.1  # via astroid
lxml==4.3.4               # via pyquery
markupsafe==1.1.1
mccabe==0.6.1             # via pylint
mock==3.0.5
oyaml==0.9
pep8==1.7.1
pip-tools==3.8.0
psycopg2==2.8.3
pyjwt==1.7.1
pylint==2.3.1
pyquery==1.4.0
python-dateutil==2.8.0    # via faker, freezegun
python-dotenv==0.10.3
pytz==2019.1
pyyaml==5.1.1
raven==6.10.0
requests==2.22.0
ruamel.yaml==0.15.97
six==1.12.0
soupsieve==1.9.2          # via beautifulsoup4
sqlparse==0.3.0
tblib==1.4.0
text-unidecode==1.2       # via faker
typed-ast==1.4.0          # via astroid
unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
uwsgi==2.0.18
vng-api-common==0.55.1
waitress==1.3.0           # via webtest
webob==1.8.5              # via webtest
webtest==2.0.33
wrapt==1.11.2             # via astroid
